массовая вставка с помощью последовательностей (#7)
* работа генераторов с nextval

* можно создавать разные последовательности для разных таблиц, новый файл миграции, отдельный метод для параллельного выполнения

* исправила имя файла миграции

* отформатировала файлы

* новый метод репозитория saveAll для массовой вставки, тест не проходит

* работающая версия saveAll (соединение не закрывается в try-with-resources)

* Не закрываю statement, убрала использование .setAccessible, метод saveAll возвращает коллекцию объектов. Черновая версия нового теста

* Добавила тест: saveAll получает одновременно и сохранённую, и не сохранённую сущности. Поменяла saveAll для поддержки insert

* реформат кода

* вынесла в отдельные методы общий код, вынесла поиск @column полей в метод

---------

+++ src/main/java/ru/inno/adeliya/jdbc/repository/generator/NextValSequenceGenerator.java
@@ -5,7 +5,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 
-public class NextValSequenceGenerator implements IdGenerator<Long> {
+public class NextValSequenceGenerator implements IdGenerator<Integer> {
     private final Connection connection;
     private final String sequenceName;
 
@@ -15,12 +15,12 @@ public class NextValSequenceGenerator implements IdGenerator<Long> {
     }
 
     @Override
-    public Long generate() {
+    public Integer generate() {
         String query = String.format("SELECT nextval('%s')", sequenceName);
         try (Statement statement = connection.createStatement();
              ResultSet resultSet = statement.executeQuery(query)) {
             if (resultSet.next()) {
-                return resultSet.getLong(1);
+                return resultSet.getInt(1);
             }
             throw new RuntimeException("No result from nextval");
         } catch (SQLException e) {
